+++ examples/nextjs/app/admin/about/page.tsx
@@ -1,0 +1,7 @@
+'use client';
+
+import { AdminAboutPage } from '@caspian-explorer/script-caspian-store';
+
+export default function AdminAboutRoute() {
+  return <AdminAboutPage />;
+}