+++ examples/nextjs/app/admin/faqs/page.tsx
@@ -1,0 +1,7 @@
+'use client';
+
+import { AdminFaqsPage } from '@caspian-explorer/script-caspian-store';
+
+export default function AdminFaqsRoute() {
+  return <AdminFaqsPage />;
+}