+++ examples/nextjs/app/admin/journal/page.tsx
@@ -1,0 +1,7 @@
+'use client';
+
+import { AdminJournalPage } from '@caspian-explorer/script-caspian-store';
+
+export default function AdminJournalRoute() {
+  return <AdminJournalPage />;
+}