+++ examples/nextjs/app/admin/promo-codes/page.tsx
@@ -1,0 +1,7 @@
+'use client';
+
+import { AdminPromoCodesPage } from '@caspian-explorer/script-caspian-store';
+
+export default function AdminPromoCodesRoute() {
+  return <AdminPromoCodesPage />;
+}