+++ examples/nextjs/app/admin/settings/[[...slug]]/page.tsx
@@ -1,0 +1,7 @@
+'use client';
+
+import { AdminSettingsShell } from '@caspian-explorer/script-caspian-store';
+
+export default function AdminSettingsRoute() {
+  return <AdminSettingsShell />;
+}