+++ examples/nextjs/app/admin/subscribers/page.tsx
@@ -1,0 +1,7 @@
+'use client';
+
+import { AdminSubscribersPage } from '@caspian-explorer/script-caspian-store';
+
+export default function AdminSubscribersRoute() {
+  return <AdminSubscribersPage />;
+}