+++ examples/nextjs/app/admin/users/page.tsx
@@ -1,0 +1,7 @@
+'use client';
+
+import { AdminUsersPage } from '@caspian-explorer/script-caspian-store';
+
+export default function AdminUsersRoute() {
+  return <AdminUsersPage />;
+}